NerbyQuests and QuestsDetails API under the Palantir controller
