+++ extensions/wikia/POI/POI.setup.php
@@ -4,7 +4,6 @@ $dir = dirname(__FILE__) . '/';
 $searchDir = dirname(dirname(__FILE__)) . '/Search/classes/Services/';
 
 //classes
-$wgAutoloadClasses['POIApiController'] =  $dir . "POIApiController.class.php";
 $wgAutoloadClasses['MetadataSpecialController'] = $dir . 'MetadataSpecialController.class.php';
 $wgAutoloadClasses['PalantirApiController'] = $dir . 'PalantirApiController.class.php';
 $wgAutoloadClasses['MetaCSVService'] = $dir . 'MetaCSVService.class.php';
