+++ extensions/wikia/POI/PalantirApiController.class.php
@@ -1,6 +1,26 @@
 <?php
 
+use Wikia\Search\Services\NearbyPOISearchService;
+
 class PalantirApiController extends WikiaApiController {
+
+	const PI = 3.14;
+
+	const EARTH_RADIUS = 6374;
+
+	const DEGREES_IN_PI = 180;
+
+	const DEFAULT_RADIUS = 180;
+
+	/**
+	 * @var QuestDetailsSolrHelper
+	 */
+	protected $solrHelper;
+
+	/**
+	 * @var Wikia\Search\Services\NearbyPOISearchService
+	 */
+	protected $nearbySearch;
 
 	/**
 	 * @var QuestDetailsSearchService
@@ -13,7 +33,7 @@ class PalantirApiController extends WikiaApiController {
 		$category = $this->getRequest()->getVal( 'category' );
 		$limit = $this->getRequest()->getVal( 'limit' );
 
-		$this->validateParameters( $limit );
+		$this->validateQuestDetailsParameters( $limit );
 
 		$result = $this->getQuestDetailsSearch()
 			->newQuery()
@@ -49,7 +69,114 @@ class PalantirApiController extends WikiaApiController {
 		return $this->questDetailsSearch;
 	}
 
-	protected function validateParameters( $limit ) {
+	protected function validateQuestDetailsParameters( $limit ) {
+		if ( !empty( $limit ) && !preg_match( '/^\d+$/i', $limit ) ) {
+			throw new BadRequestApiException( "Parameter 'limit' is invalid" );
+		}
+	}
+
+	public function getNearbyQuests() {
+		$lat = $this->getRequest()->getVal( 'latitude' );
+		$long = $this->getRequest()->getVal( 'longitude' );
+		$region = $this->getRequest()->getVal( 'region' );
+		$radius = $this->getRequest()->getVal( 'radius', self::DEFAULT_RADIUS );
+		$limit = $this->getRequest()->getVal( 'limit' );
+
+		$this->validateNearbyQuestsParameters( $lat, $long, $radius, $limit );
+
+		$radiusKilometers = $this->radiusDegreesToKilometers( $radius );
+
+		$solrHelper = $this->getSolrHelper();
+		$nearbySearch = $this->getNearbySearch();
+
+		$solrResponse = $nearbySearch->newQuery()
+			->latitude( $lat )
+			->longitude( $long )
+			->radius( $radiusKilometers )
+			->region( $region )
+			->setFields( $solrHelper->getRequiredSolrFields() )
+			->limit( $limit )
+			->withWikiaId( $this->wg->CityId )
+			->search();
+
+		$result = $solrHelper->consumeResponse( $solrResponse );
+
+		if( empty( $result ) ) {
+			throw new NotFoundApiException();
+		}
+
+		$this->setResponseData( $result );
+	}
+
+	protected function radiusDegreesToKilometers( $radiusDegrees ) {
+		return $radiusDegrees * self::EARTH_RADIUS  / self::DEGREES_IN_PI * self::PI;
+	}
+
+	/**
+	 * @param \Wikia\Search\Services\NearbyPOISearchService $nearbySearch
+	 */
+	public function setNearbySearch( $nearbySearch ) {
+		$this->nearbySearch = $nearbySearch;
+	}
+
+	/**
+	 * @return \Wikia\Search\Services\NearbyPOISearchService
+	 */
+	public function getNearbySearch() {
+		if( empty( $this->nearbySearch ) ) {
+			$this->nearbySearch = new NearbyPOISearchService();
+		}
+		return $this->nearbySearch;
+	}
+
+	/**
+	 * @param \QuestDetailsSolrHelper $solrHelper
+	 */
+	public function setSolrHelper( $solrHelper ) {
+		$this->solrHelper = $solrHelper;
+	}
+
+	/**
+	 * @return \QuestDetailsSolrHelper
+	 */
+	public function getSolrHelper() {
+		if( empty( $this->solrHelper ) ) {
+			$this->solrHelper = new QuestDetailsSolrHelper();
+		}
+		return $this->solrHelper;
+	}
+
+	protected function validateNearbyQuestsParameters( $lat, $long, $radius, $limit ) {
+		// positive and negative floating numbers
+		if ( !preg_match( '/^-?\d+(\.\d+)?$/i', $lat ) ) {
+			throw new BadRequestApiException( "Parameter 'latitude' is invalid" );
+		}
+
+		// positive and negative floating numbers
+		if ( !preg_match( '/^-?\d+(\.\d+)?$/i', $long ) ) {
+			throw new BadRequestApiException( "Parameter 'longitude' is invalid" );
+		}
+
+		$lat = doubleval( $lat );
+		if ( ( $lat < -90 ) || ( $lat > 90 ) ) {
+			throw new BadRequestApiException( "Invalid latitude: latitudes range from -90 to 90: provided latitude: ${lat}" );
+		}
+
+		$long = doubleval( $long );
+		if ( ( $long < -180 ) || ( $long > 180 ) ) {
+			throw new BadRequestApiException( "Invalid longitude: longitudes range from -180 to 180: provided longitude: ${long}" );
+		}
+
+		// only positive floating numbers
+		if ( !empty( $radius ) && !preg_match( '/^\d+(\.\d+)?$/i', $radius ) ) {
+			throw new BadRequestApiException( "Parameter 'radius' is invalid" );
+		}
+		$radius = doubleval( $radius );
+		if( $radius > 180 ) {
+			throw new BadRequestApiException( "Invalid radius: radiuses range from 0 to 180: provided radius: ${radius}" );
+		}
+
+		// only positive integer numbers
 		if ( !empty( $limit ) && !preg_match( '/^\d+$/i', $limit ) ) {
 			throw new BadRequestApiException( "Parameter 'limit' is invalid" );
 		}
